Check host key in ssh connections
Native SSH (the `ssh()` function) was covered in fab8f6c3a8e13246aaea79f266e747e2e17efa65. This covers the rest (which use the `ssh2` node library).

For reasons that are unclear to me, `ssh2` hashes the host key before passing it to the caller's `hostVerifier` function, so we need to mirror the hashing on our side in order to do the comparison. Went with `sha256` after consulting with platsec.